Rewrite of git2jss
Initial commit of the changes I've made. There are too many to enumerate in a comment, but my hope is that this will help other folks with similar issues (timeouts) as well as open the door for further refinement. It can more easily be expanded to include other types of objects like Configuration Profiles. Taking it another step and making it into a package would also be possible.